apps/analyzer/requirements.txt
fastapi
nextcord
pydantic
pymongo
python-dateutil
python-dotenv
PyYAML
requests
uvicorn
transformers
sentencepiece
protobuf